pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexxy32"
version = "0.1.0"
requires-python = ">= 3.12"
dependencies = [
    "pandas>=2.2.3",
    "riscof @ git+https://github.com/riscv/riscof@9fe3597d75",
    # https://github.com/riscv-software-src/riscof/issues/122
    "riscv-isac @ git+https://github.com/riscv-software-src/riscv-isac@777d2b4762",
    "fastapi>=0.115.5",
    "websockets>=14.1",
    "uvicorn>=0.32.1",
    "typer>=0.13.1",
    "httpx>=0.27.2",
    "opencv-python>=4.10.0.84",
    "tqdm>=4.67.1",
    "arduino-simple-rpc>=2.4.2",
]

[tool.rye]
managed = true
dev-dependencies = [
    "ruff~=0.7.1",
    "ipykernel>=6.29.5",
]

[tool.rye.scripts]
server = { call = "hexxy32.server.app" }
hexxy32 = { call = "hexxy32.cli.app" }

[tool.hatch.build.targets.wheel]
packages = ["python/hexxy32"]

[tool.hatch.metadata]
allow-direct-references = true

# linting/formatting

[tool.ruff]
preview = true
extend-exclude = [
    "**/venv",
    "**/.*",
    "**/__pycache__",
    "**/build",
    "**/target",
]

[tool.ruff.lint]
extend-select = [
    "I", # import sorting
    "RUF022", # __all__ sorting
]
extend-ignore = [
    # covered by Pyright
    "F821", # undefined variable
    "F401", # imported but unused
    "F841", # unused variable
]

[tool.ruff.lint.isort]
combine-as-imports = true

# type checking

[tool.pyright]
venvPath = "."
venv = ".venv"

include = [
    "src",
]
exclude = [
    "venv",
    ".venv",
    "build",
    "target",
    "riscof/sail_cSim",
]

typeCheckingMode = "basic"

enableExperimentalFeatures = true
strictDictionaryInference = true
strictListInference = true
strictSetInference = true

reportAssertAlwaysTrue = "error"
reportConstantRedefinition = "error"
reportDeprecated = "error"
reportFunctionMemberAccess = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportIncompleteStub = "error"
reportInconsistentConstructor = "error"
reportInvalidStringEscapeSequence = "error"
reportInvalidStubStatement = "error"
reportInvalidTypeVarUse = "error"
reportMatchNotExhaustive = "error"
reportMissingParameterType = "error"
reportOverlappingOverload = "error"
reportSelfClsParameterName = "error"
reportTypeCommentUsage = "error"
reportUnknownParameterType = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryContains = "error"
reportUnsupportedDunderAll = "error"
reportUntypedBaseClass = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedNamedTuple = "error"
reportWildcardImportFromLibrary = "error"

reportMissingTypeArgument = "warning"
reportPrivateUsage = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnusedClass = "warning"
reportUnusedExpression = "warning"
reportUnusedFunction = "warning"
reportUnusedImport = "warning"
reportUnusedVariable = "warning"

reportMissingTypeStubs = "none"
reportDuplicateImport = "none"
reportUnknownArgumentType = "none"
reportUnknownLambdaType = "none"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
